+++ packages/swc-plugin-extractor/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 4.10.0 (2026-04-28)
+
+### Features
+
+* Add per-domain `localePrefix` override support ([#2273](https://github.com/amannn/next-intl/issues/2273)) ([3e9febf](https://github.com/amannn/next-intl/commit/3e9febfc14f26d3f13ae0c20059c78028c22d50d)) – by @frankmatheron
+
+## 4.9.2 (2026-04-27)
+
+### Bug Fixes
+
+* Prototype safety guards for `precompile: true` ([#2307](https://github.com/amannn/next-intl/issues/2307)) ([c0bf0ee](https://github.com/amannn/next-intl/commit/c0bf0ee42abb38926dee51474a80b9982b3acb2f)) – by @amannn
+
+## 4.9.1 (2026-04-10)
+
+### Bug Fixes
+
+* Improve middleware pathname validation ([#2304](https://github.com/amannn/next-intl/issues/2304)) ([1c80b66](https://github.com/amannn/next-intl/commit/1c80b668aa6d853f470319eec10a3f61e78a70e6)) – by @amannn
+
+## 4.9.0 (2026-04-01)
+
+### Features
+
+* Support `transitionTypes` on `Link` ([#2302](https://github.com/amannn/next-intl/issues/2302)) ([02811f5](https://github.com/amannn/next-intl/commit/02811f5dd5f31753d742632082874b6e3e3667f6)) – by @amannn
+
+## 4.8.4 (2026-03-31)
+
+### Bug Fixes
+
+* Remove TypeScript peer dependency and update examples to TypeScript v6 ([#2293](https://github.com/amannn/next-intl/issues/2293)) ([5e7bcd7](https://github.com/amannn/next-intl/commit/5e7bcd743994f8a3ccfd904d7969f4543950cd0a)) – by @wojtekmaj
+
 ## 4.8.3 (2026-02-16)
 
 ### Bug Fixes